backend/requirements.txt
fastapi>=0.109.0,<0.110.0
uvicorn[standard]>=0.27.0,<0.28.0
python-multipart>=0.0.6,<0.0.7
pydantic>=2.6.0,<3.0.0
anthropic>=0.19.0,<0.20.0
python-dotenv>=1.0.0,<2.0.0
